initial refacctoring to work with alchemy API and generally to support sqrtPriceX96 fields as the main difference.
